% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_quotes_if_has_spaces.R
\name{add_quotes_if_has_spaces}
\alias{add_quotes_if_has_spaces}
\title{Add quotes around the string if it contains spaces.}
\usage{
add_quotes_if_has_spaces(filename)
}
\arguments{
\item{filename}{a filename}
}
\value{
a filename. If the filename did not contain spaces,
it is returned as-is. If the filename did contain spaces,
the filename is surrounded by quotes
}
\description{
Add quotes around the string if it contains spaces.
Does nothing if the string contains no spaces.
This is used for filenames
}
