% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_identical_alignments.R
\name{are_identical_alignments}
\alias{are_identical_alignments}
\title{Determines if the two alignments are equal}
\usage{
are_identical_alignments(p, q)
}
\arguments{
\item{p}{the first alignment}

\item{q}{the second alignment}
}
\value{
TRUE or FALSE
}
\description{
Determines if the two alignments are equal
}
\author{
Richel Bilderbeek
}
