% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast2_internal_filenames_to_table.R
\name{beast2_internal_filenames_to_table}
\alias{beast2_internal_filenames_to_table}
\title{Convert a \code{beast2_internal_filenames} to a table}
\usage{
beast2_internal_filenames_to_table(beast2_internal_filenames)
}
\arguments{
\item{beast2_internal_filenames}{a list of internally
used BEAST2 filenames,
as created by \link{create_beast2_internal_filenames}}
}
\description{
Convert a \code{beast2_internal_filenames} to a table
}
