% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastier.R
\docType{package}
\name{beastier}
\alias{beastier}
\alias{beastier-package}
\title{beastier: A package to call BEAST2.}
\description{
'beastier' allows to call BEAST2, a popular
Bayesian phylogenetics tool, using
an R interface. 'beastier' closely follows the interface
of BEAST2, including its default settings.
}
\seealso{
The package \link[beautier]{beautier} can create
  BEAST2 input files from R.
}
