% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_internal_filenames.R
\name{check_beast2_internal_filenames_names}
\alias{check_beast2_internal_filenames_names}
\title{Check if the \code{beast2_internal_filenames}, which is a list,
has all the elements needed.}
\usage{
check_beast2_internal_filenames_names(beast2_internal_filenames)
}
\arguments{
\item{beast2_internal_filenames}{a list of internally
used BEAST2 filenames,
as created by \link{create_beast2_internal_filenames}}
}
\value{
nothing
}
\description{
Calls \code{stop} if not.
}
\seealso{
Use \link{check_beast2_internal_filenames} to check
  the entire \code{beast2_internal_filenames} object
}
\author{
Richèl J.C. Bilderbeek
}
