% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_jar_path.R
\name{check_beast2_jar_path}
\alias{check_beast2_jar_path}
\title{Checks the BEAST2 \code{.jar} path.
Will stop if there is a problem with the BEAST2 \code{.jar} path.}
\usage{
check_beast2_jar_path(beast2_jar_path)
}
\arguments{
\item{beast2_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
nothing.
  Will call \code{\link{stop}} if the BEAST2 \code{.jar} path has a problem
}
\description{
Checks the BEAST2 \code{.jar} path.
Will stop if there is a problem with the BEAST2 \code{.jar} path.
}
\author{
Richel J.C. Bilderbeek
}
