% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_optionses.R
\name{check_beast2_optionses}
\alias{check_beast2_optionses}
\title{Check if the \code{beast2_options} is a valid BEAST2 options object.}
\usage{
check_beast2_optionses(beast2_optionses)
}
\value{
nothing
}
\description{
Calls \code{stop} if the BEAST2 option object is invalid
}
\examples{
 testthat::expect_silent(
   check_beast2_optionses(
       list(create_beast2_options()
     )
   )
 )

 # Must stop on nonsense
 testthat::expect_error(check_beast2_optionses("nonsense"))
 testthat::expect_error(check_beast2_optionses(NULL))
 testthat::expect_error(check_beast2_optionses(NA))
}
\seealso{
Use \link{create_beast2_options} to create a valid
  BEAST2 options object
}
\author{
Richel J.C. Bilderbeek
}
