% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_can_create_file.R
\name{check_can_create_file}
\alias{check_can_create_file}
\title{Check that a file can be created at a certain path.}
\usage{
check_can_create_file(filename)
}
\arguments{
\item{filename}{file that may or may not be created}
}
\description{
Will \link{stop} if not. Will \link{stop} if the file already exists.
Does so by creating an empty file at the path,
and then deleting it.
}
\author{
Richèl J.C. Bilderbeek
}
