% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_can_create_tracelog_file.R
\name{check_can_create_tracelog_file}
\alias{check_can_create_tracelog_file}
\title{Internal function to check if the MCMC's tracelog file can be created.}
\usage{
check_can_create_tracelog_file(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\description{
Check if the MCMC's tracelog file can be created.
Will \link{stop} if not.
If the tracelog file already exists,
it is assumed that a new file can be created,
by overwriting the existing one.
}
\author{
Richèl J.C. Bilderbeek
}
