% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_beastier_folder.R
\name{check_empty_beastier_folder}
\alias{check_empty_beastier_folder}
\title{Check there are no files in the default \link{beastier} folder}
\usage{
check_empty_beastier_folder(beastier_folder = get_beastier_folder())
}
\value{
Nothing.
}
\description{
Check there are no files in the default \link{beastier} folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the \link{beastier} folder
}
\author{
Richèl J.C. Bilderbeek
}
