% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_beaustier_folders.R
\name{check_empty_beaustier_folders}
\alias{check_empty_beaustier_folders}
\title{Check there are no files in the default
`beautier` and `beastier` folders}
\usage{
check_empty_beaustier_folders(
  beautier_folder = beautier::get_beautier_folder(),
  beastier_folder = beastier::get_beastier_folder()
)
}
\arguments{
\item{beautier_folder}{temporary folder used by \link[beautier]{beautier}}

\item{beastier_folder}{the path to
the \link{beastier} temporary files folder}
}
\value{
Nothing.
Will \link{stop} if there are files in the \link[beautier]{beautier}
of \link{beastier} folder.
}
\description{
Check there are no files in the default
`beautier` and `beastier` folders.
}
\details{
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the \link[beautier]{beautier}
of \link{beastier} folder.
}
\author{
Richèl J.C. Bilderbeek
}
