% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_filename.R
\name{check_input_filename}
\alias{check_input_filename}
\title{Checks the input filename.
Will stop if there is a problem with the input filename.}
\usage{
check_input_filename(input_filename)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.}
}
\value{
nothing. Will call \code{\link{stop}} if the input file is invalid
}
\description{
Checks the input filename.
Will stop if there is a problem with the input filename.
}
\author{
Richèl J.C. Bilderbeek
}
