% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_beast2_working_dir.R
\name{clear_beast2_working_dir}
\alias{clear_beast2_working_dir}
\title{Clear the BEAST2 working directory}
\usage{

  clear_beast2_working_dir(beast2_working_dir = get_default_beast2_working_dir())
}
\description{
To assure one cannot accidentally delete important folders, 
there are some checks on the name of the supplied \code{beast2_working_dir}
}
