% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_beastier_cache.R
\name{clear_beastier_cache}
\alias{clear_beastier_cache}
\title{Clear the \link{beastier} cache}
\usage{
clear_beastier_cache(beastier_folder = get_beastier_folder())
}
\value{
Nothing.
}
\description{
Clear the \link{beastier} cache.
}
\author{
Richèl J.C. Bilderbeek
}
