% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_cmd_validate.R
\name{create_beast2_cmd_validate}
\alias{create_beast2_cmd_validate}
\title{Creates the terminal command to validate a BEAST2 input file}
\usage{
create_beast2_cmd_validate(input_filename,
  beast_jar_path = "~/Programs/beast/lib/beast.jar")
}
\arguments{
\item{input_filename}{name of the BEAST2 input file
(usually has a \code{.xml} extension)}

\item{beast_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension)}
}
\description{
Creates the terminal command to validate a BEAST2 input file
}
\author{
Richel J.C. Bilderbeek
}
