% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_options.R
\name{create_beast2_options}
\alias{create_beast2_options}
\title{Function to create a set of BEAST2 options.}
\usage{
create_beast2_options(input_filename = tempfile(fileext = ".xml"),
  output_log_filename = tempfile(fileext = ".log"),
  output_trees_filenames = tempfile(fileext = ".trees"),
  output_state_filename = tempfile(fileext = ".xml.state"),
  rng_seed = NA, n_threads = NA, use_beagle = FALSE,
  overwrite = TRUE,
  beast2_working_dir = file.path(get_default_beast2_working_dir(),
  basename(tempfile())), beast2_path = get_default_beast2_path(),
  verbose = FALSE)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.}

\item{output_log_filename}{name of the .log file to create}

\item{output_trees_filenames}{one or more names for .trees file to create.
There will be one .trees file created per alignment in the input
file. The number of alignments must equal the number of .trees
filenames, else an error is thrown. Alignments are sorted alphabetically
by their IDs}

\item{output_state_filename}{name of the .xml.state file to create}

\item{rng_seed}{the random number generator seed of the BEAST2 run.
Must be a non-zero positive integer value or \link{NA}.
If \code{rng_seed} is \link{NA}, BEAST2 will pick a random seed}

\item{n_threads}{the number of computational threads to use.
Use \link{NA} to use the BEAST2 default of 1.}

\item{use_beagle}{use BEAGLE if present}

\item{overwrite}{if TRUE: overwrite the \code{.log}
 and \code{.trees} files if one of these exists.
 If FALSE, BEAST2 will not be started if
 \itemize{
   \item{the \code{.log} file exists}
   \item{the \code{.trees} files exist}
   \item{the \code{.log} file created by BEAST2 exists}
   \item{the \code{.trees} files created by BEAST2 exist}
}}

\item{beast2_working_dir}{a (temporary) folder where BEAST2 can work in
isolation.  This is a new and empty temporary folder by default.
In this folder, BEAST2 is allowed to create all of its output files,
without the risk of overwriting existing ones.
Currently, it appears possible that one can set
\code{beast2_working_dir} to the same longer-storage folder
as where the BEAST2 output files are created, but this may change
in future version.
Using this working directory allows to call
BEAST2 in multiple parallel processes, as each process can have
its own separate environment}

\item{beast2_path}{name of either a BEAST2 binary file
(usually simply \code{beast})
or a BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_bin_path} to get
the default BEAST binary file's path
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\value{
a BEAST2 options structure
}
\description{
These BEAST2 options are the R equivalent of the command-line options.
}
\examples{
  library(testthat)

  beast2_options <- create_beast2_options()

  expect_true("input_filename" \%in\% names(beast2_options))
  expect_true("output_log_filename" \%in\% names(beast2_options))
  expect_true("output_trees_filenames" \%in\% names(beast2_options))
  expect_true("output_state_filename" \%in\% names(beast2_options))
  expect_true("rng_seed" \%in\% names(beast2_options))
  expect_true("n_threads" \%in\% names(beast2_options))
  expect_true("use_beagle" \%in\% names(beast2_options))
  expect_true("overwrite" \%in\% names(beast2_options))
  expect_true("beast2_path" \%in\% names(beast2_options))
  expect_true("verbose" \%in\% names(beast2_options))

  expect_silent(check_beast2_options(beast2_options))
}
\author{
Richèl J.C. Bilderbeek
}
