% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_run_cmd.R
\name{create_beast2_run_cmd}
\alias{create_beast2_run_cmd}
\title{Creates the terminal command to run BEAST2}
\usage{
create_beast2_run_cmd(input_filename, output_state_filename, rng_seed = NA,
  n_threads = NA, use_beagle = FALSE, overwrite_state_file = TRUE,
  beast_jar_path = "~/Programs/beast/lib/beast.jar")
}
\arguments{
\item{input_filename}{name of the BEAST2 input file
(usually has a \code{.xml} extension)}

\item{output_state_filename}{name of the BEAST2 output file that
stores the state
(usually has a \code{.xml.state} extension)}

\item{rng_seed}{the RNG seed}

\item{n_threads}{number of threads to use}

\item{use_beagle}{use BEAGLE if present}

\item{overwrite_state_file}{set to TRUE to overwrite the file with name
\code{output_state_filename} if it already exists}

\item{beast_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension)}
}
\description{
Creates the terminal command to run BEAST2
}
\author{
Richel J.C. Bilderbeek
}
