% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_screenlog_folder.R
\name{create_beast2_screenlog_folder}
\alias{create_beast2_screenlog_folder}
\title{Internal function}
\usage{
create_beast2_screenlog_folder(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\description{
Create the folder for the BEAST2 screenlog file
}
\author{
Richèl J.C. Bilderbeek
}
