% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_state_output_file_folder.R
\name{create_beast2_state_output_file_folder}
\alias{create_beast2_state_output_file_folder}
\title{Create the folder where the BEAST2 state output file will be created}
\usage{
create_beast2_state_output_file_folder(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\value{
nothing
}
\description{
Create the folder where the BEAST2 state output file will be created
}
\examples{
beast2_options <- create_beast2_options()
create_beast2_state_output_file_folder()
}
\author{
Richèl J.C. Bilderbeek
}
