% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_validate_cmd.R
\name{create_beast2_validate_cmd}
\alias{create_beast2_validate_cmd}
\title{Creates the terminal command to validate a BEAST2 input file}
\usage{
create_beast2_validate_cmd(input_filename,
  beast2_jar_path = get_default_beast2_jar_path())
}
\arguments{
\item{input_filename}{name of the BEAST2 input file
(usually has a \code{.xml} extension)}

\item{beast2_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\description{
Creates the terminal command to validate a BEAST2 input file
}
\author{
Richel J.C. Bilderbeek
}
