% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_log_filename.R
\name{create_default_log_filename}
\alias{create_default_log_filename}
\title{Create the default \code{.log} filename}
\usage{
create_default_log_filename(input_filename, beast2_jar_path, verbose)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file}
}
\description{
Create the default \code{.log} filename
}
\author{
: Richel J.C. Bilderbeek
}
