% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_trees_filename.R
\name{create_default_trees_filenames}
\alias{create_default_trees_filenames}
\title{Create the default \code{.trees} filenames}
\usage{
create_default_trees_filenames(input_filename, beast2_jar_path, verbose)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file}
}
\description{
Create the default \code{.trees} filenames
}
\author{
: Richel J.C. Bilderbeek
}
