% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_posterior.R
\name{create_posterior}
\alias{create_posterior}
\title{Create a BEAST2 posterior by running BEAST2 from a random alignment}
\usage{
create_posterior(n_taxa, sequence_length, mcmc, crown_ages = NA,
  verbose = FALSE)
}
\arguments{
\item{n_taxa}{number of taxa in the simulated phylogeny}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{mcmc}{one \code{beautier} mcmc object}

\item{crown_ages}{the crown ages of the phylogenies. Set to NA
if the crown age needs to be estimated}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
Create a BEAST2 posterior by running BEAST2 from a random alignment
}
\author{
Richel J.C. Bilderbeek
}
