% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_alignment.R
\name{create_random_alignment}
\alias{create_random_alignment}
\title{Create a random alignment}
\usage{
create_random_alignment(n_taxa, sequence_length, rate = 1,
  taxa_name_ext = "")
}
\arguments{
\item{n_taxa}{The number of taxa}

\item{sequence_length}{The number of base pairs the alignment will have}

\item{rate}{mutation rate}

\item{taxa_name_ext}{the extension of the taxa names}
}
\value{
an alignment of class DNAbin
}
\description{
Create a random alignment
}
\examples{
 alignment <- create_random_alignment(
   n_taxa = 5,
   sequence_length = 10
 )
 image(alignment)
}
\author{
Richel J.C. Bilderbeek
}
