% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_fastas.R
\name{create_random_fastas}
\alias{create_random_fastas}
\title{Create one or more random FASTA files}
\usage{
create_random_fastas(n_taxa, sequence_length, fasta_filenames)
}
\arguments{
\item{n_taxa}{The number of taxa}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{fasta_filenames}{One or more FASTA filenames.}
}
\value{
Nothing, creates one or more random FASTA files
}
\description{
Create one or more random FASTA files
}
\author{
Richel J.C. Bilderbeek
}
