% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_random_phylogeny.R
\name{create_random_phylogeny}
\alias{create_random_phylogeny}
\title{Create a random phylogeny}
\usage{
create_random_phylogeny(n_taxa = n_taxa, taxa_name_ext = "")
}
\arguments{
\item{n_taxa}{The number of taxa}

\item{taxa_name_ext}{the extension of the taxa names}
}
\value{
a phylogeny of class \code{phylo}
}
\description{
Create a random phylogeny
}
\author{
Richel J.C. Bilderbeek
}
