% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_minimal_run.R
\name{do_minimal_run}
\alias{do_minimal_run}
\title{Do a minimal BEAST2 run}
\usage{
do_minimal_run()
}
\value{
The text sent to \code{STDOUT} and \code{STDERR}.
  It will create the files with names
  \code{output_log_filename}, \code{output_trees_filenames}
  and \code{output_state_filenames}
}
\description{
To achieve this, \link{run_beast2_from_options} is called.
}
\examples{
  library(testthat)

  if (is_beast2_installed() && is_on_ci()) {

    expect_silent(do_minimal_run())

  }
}
\author{
Richèl J.C. Bilderbeek
}
