% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/extract_tracelog_filename_from_beast2_input_file.R
\name{extract_tracelog_filename_from_beast2_input_file}
\alias{extract_tracelog_filename_from_beast2_input_file}
\title{Internal function to extract the tracelog filename for a BEAST2 input file}
\usage{
extract_tracelog_filename_from_beast2_input_file(input_filename)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.
Use \link{create_temp_input_filename} to create a temporary
filename with that extension.}
}
\description{
Extract the tracelog filename for a BEAST2 input file
}
\author{
Richèl J.C. Bilderbeek
}
