% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_example_filename.R
\name{get_beast2_example_filename}
\alias{get_beast2_example_filename}
\title{Get the full path of a BEAST2 example file}
\usage{
get_beast2_example_filename(filename,
  folder = get_default_beast2_folder())
}
\description{
Will \link{stop} if the filename is not a BEAST2 example file
}
