% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_example_filenames.R
\name{get_beast2_example_filenames}
\alias{get_beast2_example_filenames}
\title{Get a list with the full paths of all
BEAST2 example filenames}
\usage{

  get_beast2_example_filenames(beast2_folder = get_default_beast2_folder())
}
\value{
a list with the full paths of all
BEAST2 example filenames
}
\description{
Get a list with the full paths of all
BEAST2 example filenames
}
