% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_main_class_name.R
\name{get_beast2_main_class_name}
\alias{get_beast2_main_class_name}
\title{Get the BEAST2 main class name.}
\usage{
get_beast2_main_class_name()
}
\description{
One way to fix the error
\code{no main manifest attribute}
is to specify the main class name.
}
