% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_options_filenames.R
\name{get_beast2_options_filenames}
\alias{get_beast2_options_filenames}
\title{Extract the filenames from a \code{beast2_options}}
\usage{
get_beast2_options_filenames(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\description{
Extract the filenames from a \code{beast2_options}
}
\examples{
library(testthat)

beast2_options <- beastier::create_beast2_options()
filenames <- get_beast2_options_filenames(beast2_options)
expect_true(beast2_options$input_filename \%in\% filenames)
expect_true(beast2_options$output_state_filename \%in\% filenames)
}
\author{
Richèl J.C. Bilderbeek
}
