% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_version.R
\name{get_beast2_version}
\alias{get_beast2_version}
\title{Get the BEAST2 version}
\usage{
get_beast2_version(beast2_path = get_default_beast2_jar_path())
}
\description{
Get the BEAST2 version
}
\examples{
library(testthat)

if (is_beast2_installed() && is_on_ci()) {
  expect_equal(get_beast2_version(), "2.6.0")
}
}
\author{
Richèl J.C. Bilderbeek
}
