% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beastier_folder.R
\name{get_beastier_folder}
\alias{get_beastier_folder}
\title{Get the path to the \link{beastier} temporary files folder}
\usage{
get_beastier_folder()
}
\value{
the path to the \link{beastier} temporary files folder.
}
\description{
Get the path to the \link{beastier} temporary files folder.
}
\examples{
get_beastier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
