% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path.R
\name{get_beastier_path}
\alias{get_beastier_path}
\title{Get the full path of a file in the \code{inst/extdata} folder}
\usage{
get_beastier_path(filename)
}
\arguments{
\item{filename}{the file's name, without the path}
}
\value{
the full path to the filename
}
\description{
Get the full path of a file in the \code{inst/extdata} folder
}
\examples{
  testit::assert(
    is.character(get_beastier_path("beast2_example_output.log"))
  )
  testit::assert(
    is.character(get_beastier_path("beast2_example_output.trees"))
  )
  testit::assert(
    is.character(get_beastier_path("beast2_example_output.xml"))
  )
  testit::assert(
    is.character(get_beastier_path("beast2_example_output.xml.state"))
  )
}
\seealso{
for more files, use \code{\link{get_beastier_paths}}
}
\author{
Richel J.C. Bilderbeek
}
