% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_paths.R
\name{get_beastier_paths}
\alias{get_beastier_paths}
\title{Get the full paths of files in the \code{inst/extdata} folder}
\usage{
get_beastier_paths(filenames)
}
\arguments{
\item{filenames}{the files' names, without the path}
}
\value{
the filenames' full paths. Will \code{stop} if a file
  is absent in the \code{inst/extdata} folder
}
\description{
Get the full paths of files in the \code{inst/extdata} folder
}
\examples{
  library(testthat)

  filenames <- get_beastier_paths(
    c(
      "beast2_example_output.log",
      "beast2_example_output.trees",
      "beast2_example_output.xml",
      "beast2_example_output.xml.state"
    )
  )

  expect_equal(length(filenames), 4)
  expect_true(files_exist(filenames))
}
\seealso{
for one file, use \code{\link{get_beastier_path}}
}
\author{
Richèl J.C. Bilderbeek
}
