% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_bin_path.R
\name{get_default_beast2_bin_path}
\alias{get_default_beast2_bin_path}
\title{Get the default BEAST2 binary file (\code{beast}, that is) path}
\usage{
get_default_beast2_bin_path(os = rappdirs::app_dir()$os)
}
\value{
the default BEAST2 binary file's path
}
\description{
Get the default BEAST2 binary file (\code{beast}, that is) path
}
\examples{
  if (rappdirs::app_dir()$os == "unix") {
    testit::assert(
      grepl(
        "/home/[A-Za-z0-9_]*/.local/share/beast/bin/beast",
        get_default_beast2_bin_path()
      )
    )
  }
}
\seealso{
Use \link{get_default_beast2_folder} to get the default
  folder in which BEAST2 is installed.
  Use \link{install_beast2} with default arguments
  to install BEAST2 to this location.
}
\author{
Richel J.C. Bilderbeek
}
