% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_download_url.R
\name{get_default_beast2_download_url_win}
\alias{get_default_beast2_download_url_win}
\title{Get the BEAST2 download URL for Windows}
\usage{
get_default_beast2_download_url_win()
}
\value{
the URL where BEAST2 can be downloaded from
}
\description{
Get the BEAST2 download URL for Windows
}
\author{
Richèl J.C. Bilderbeek
}
