% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_version.R
\name{get_default_beast2_version}
\alias{get_default_beast2_version}
\title{Get the default BEAST2 version that is used by beastier}
\usage{
get_default_beast2_version()
}
\value{
the BEAST2 version
}
\description{
Get the default BEAST2 version that is used by beastier
}
\examples{
check_empty_beaustier_folders()

get_default_beast2_version()

check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
