% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_working_dir.R
\name{get_default_beast2_working_dir}
\alias{get_default_beast2_working_dir}
\title{Get the path to the default BEAST2 working directory}
\usage{
get_default_beast2_working_dir()
}
\description{
Get the path to the default BEAST2 working directory
}
