% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_java_path.R
\name{get_default_java_path}
\alias{get_default_java_path}
\title{Obtains the default path to the Java executable}
\usage{
get_default_java_path()
}
\value{
the default path to the Java executable
}
\description{
Obtains the default path to the Java executable
}
\author{
Richel J.C. Bilderbeek
}
