% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_java_version.R
\name{get_java_version}
\alias{get_java_version}
\title{Get the Java version}
\usage{
get_java_version()
}
\description{
Get the Java version
}
\examples{

if (is_beast2_installed() && is_on_ci()) {
  expect_silent(get_java_version())
}
}
\author{
Richèl J.C. Bilderbeek
}
