% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gives_beast2_warning.R
\name{gives_beast2_warning}
\alias{gives_beast2_warning}
\title{Determines if BEAST2 issues a warning
when using the BEAST2 XML input file}
\usage{
gives_beast2_warning(filename, verbose = FALSE,
  beast2_jar_path = get_default_beast2_jar_path())
}
\arguments{
\item{filename}{name of the BEAST2 XML input file}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}

\item{beast2_jar_path}{the path of \code{beast.jar}.
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
TRUE if the file produces a BEAST2 warning, FALSE if not
}
\description{
Determines if BEAST2 issues a warning
when using the BEAST2 XML input file
}
\seealso{
Use \code{\link{are_beast2_input_lines}} to check the lines
}
\author{
Richel J.C. Bilderbeek
}
