% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_alignment.R
\name{is_alignment}
\alias{is_alignment}
\title{Determines if the input is an alignment of type \link[ape]{DNAbin}}
\usage{
is_alignment(input)
}
\arguments{
\item{input}{The input to be tested}
}
\value{
TRUE or FALSE
}
\description{
Determines if the input is an alignment of type \link[ape]{DNAbin}
}
\author{
Richèl J.C. Bilderbeek
}
