% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_jar_path.R
\name{is_jar_path}
\alias{is_jar_path}
\title{Is the path a path to the BEAST2 jar file?
Does not check if the file at that path is present}
\usage{
is_jar_path(path)
}
\arguments{
\item{path}{a string to a path}
}
\value{
TRUE if the path is a path to a BEAST2 jar file
}
\description{
Is the path a path to the BEAST2 jar file?
Does not check if the file at that path is present
}
\examples{
# Returns TRUE
is_jar_path("beast.jar")
is_jar_path("launcher.jar")
is_jar_path(get_default_beast2_jar_path())
# Returns FALSE
is_jar_path("beast")
is_jar_path(get_default_beast2_bin_path())
}
\author{
Richèl J.C. Bilderbeek
}
