% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_appveyor}
\alias{is_on_appveyor}
\title{Deprecated function, use \link[beautier]{is_on_appveyor}}
\usage{
is_on_appveyor()
}
\value{
TRUE if run on AppVeyor, FALSE otherwise
}
\description{
Deprecated function, use \link[beautier]{is_on_appveyor}
}
\author{
Richèl J.C. Bilderbeek
}
