% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_ci}
\alias{is_on_ci}
\title{Deprecated function, use \link[beautier]{is_on_ci}}
\usage{
is_on_ci()
}
\value{
TRUE if run on AppVeyor or Travis CI, FALSE otherwise
}
\description{
Deprecated function, use \link[beautier]{is_on_ci}
}
\author{
Richèl J.C. Bilderbeek
}
