% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_travis}
\alias{is_on_travis}
\title{Deprecated function, use \link[beautier]{is_on_travis}}
\usage{
is_on_travis()
}
\value{
TRUE if run on Travis CI, FALSE otherwise
}
\description{
Deprecated function, use \link[beautier]{is_on_travis}
}
\author{
Richèl J.C. Bilderbeek
}
