% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beast2_internal_filenames.R
\name{print_beast2_internal_filenames}
\alias{print_beast2_internal_filenames}
\title{Pretty-print a \code{beast2_options}}
\usage{
print_beast2_internal_filenames(beast2_internal_filenames,
  show_exist = FALSE)
}
\description{
Pretty-print a \code{beast2_options}
}
