% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_beastier_folder.R
\name{remove_beastier_folder}
\alias{remove_beastier_folder}
\title{Check there are no files in the default \link[beautier]{beautier} folder}
\usage{
remove_beastier_folder()
}
\value{
Nothing.
}
\description{
Check there are no files in the default \link[beautier]{beautier} folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the \link[beautier]{beautier} folder.
}
\examples{
check_empty_beastier_folder()

remove_beastier_folder()

check_empty_beastier_folder()
}
\seealso{
use \link[beautier]{remove_beautier_folder} to remove the default
`beautier` folder
}
\author{
Richèl J.C. Bilderbeek
}
