% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_beaustier_folders.R
\name{remove_beaustier_folders}
\alias{remove_beaustier_folders}
\title{Remove the `beautier` and `beastier` temporary folders}
\usage{
remove_beaustier_folders()
}
\value{
Nothing.
}
\description{
Remove the `beautier` and `beastier` temporary folders
}
\examples{
check_empty_beaustier_folders()

remove_beaustier_folders()

check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
