% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_beautier_folder.R
\name{remove_beautier_folder}
\alias{remove_beautier_folder}
\title{Check there are no files in the default `beautier` folder}
\usage{
remove_beautier_folder()
}
\value{
Nothing.
}
\description{
Check there are no files in the default `beautier` folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the `beautier` folder.
}
\examples{
remove_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
