% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_file_if_present.R
\name{remove_file_if_present}
\alias{remove_file_if_present}
\title{Remove a file if it is present}
\usage{
remove_file_if_present(filename)
}
\arguments{
\item{filename}{name of a file}
}
\description{
Remove a file if it is present
}
