% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_files.R
\name{remove_files}
\alias{remove_files}
\title{Removes files if and only if present}
\usage{
remove_files(filenames)
}
\arguments{
\item{filenames}{names of the files to be removed}
}
\value{
nothing, removes those files that are present
}
\description{
Removes files if and only if present
}
\examples{
  # Can safely delete absent files
  beastier:::remove_files("abs.ent")

  # Create a file
  filename <- "pres.ent"
  file.copy(from = get_path("anthus_2_4.xml"), to = filename)
  testit::assert(file.exists(filename))

  # Can safely delete a mix of present and absent files
  beastier:::remove_files(c("abs.ent", filename))
  testit::assert(!file.exists(filename))
}
\author{
Richel J.C. Bilderbeek
}
