% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_file_if_present.R
\name{remove_files_if_present}
\alias{remove_files_if_present}
\title{Remove a file if it is present}
\usage{
remove_files_if_present(filenames)
}
\arguments{
\item{filename}{one or more names of files}
}
\description{
Remove a file if it is present
}
