% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_beast2.R
\name{run_beast2}
\alias{run_beast2}
\title{Run BEAST2}
\usage{
run_beast2(input_filename, output_log_filename = "out.log",
  output_trees_filenames = "out.trees",
  output_state_filename = "out.xml.state",
  beast_jar_path = "~/Programs/beast/lib/beast.jar", verbose = FALSE)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file}

\item{beast_jar_path}{the path of \code{beast.jar}}

\item{verbose}{show more (debug) output}
}
\description{
Run BEAST2
}
\author{
Richel J.C. Bilderbeek
}
