% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_nexus_as_fasta.R
\name{save_nexus_as_fasta}
\alias{save_nexus_as_fasta}
\title{Save a NEXUS file as a FASTA file}
\usage{
save_nexus_as_fasta(nexus_filename, fasta_filename)
}
\arguments{
\item{nexus_filename}{name of an existing NEXUS file}

\item{fasta_filename}{name of the FASTA file to be created}
}
\description{
Save a NEXUS file as a FASTA file
}
