% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_beastier.R
\name{update_beastier}
\alias{update_beastier}
\title{Update all beastier dependencies, by installing their
latest versions}
\usage{
update_beastier()
}
\description{
Update all beastier dependencies, by installing their
latest versions
}
\author{
Richèl J.C. Bilderbeek
}
