% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_beast2.R
\name{upgrade_beast2}
\alias{upgrade_beast2}
\title{Upgrade BEAST2.}
\usage{
upgrade_beast2(folder_name = rappdirs::user_data_dir(),
  os = rappdirs::app_dir()$os)
}
\arguments{
\item{folder_name}{name of the folder where the BEAST2 files will
be put.
The name of the BEAST2 binary file will be at
\code{[folder_name]/beast/bin/beast}
The name of the BEAST2 jar file will be at
\code{[folder_name]/beast/lib/launcher.jar}}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\description{
Will \link{stop} if BEAST2 is not installed
}
\examples{
library(testthat)

if (is_beast2_installed() && is_on_ci()) {
  expect_equal(get_beast2_version(), "2.6.0")
}
}
\author{
Richèl J.C. Bilderbeek
}
