% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast2_options_to_table.R
\name{beast2_options_to_table}
\alias{beast2_options_to_table}
\title{Convert a \code{beast2_options} to a table}
\usage{
beast2_options_to_table(beast2_options)
}
\description{
Convert a \code{beast2_options} to a table
}
