% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beast2_options_to_table.R
\name{beast2_options_to_table}
\alias{beast2_options_to_table}
\title{Convert a \code{beast2_options} to a table}
\usage{
beast2_options_to_table(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\description{
Convert a \code{beast2_options} to a table
}
