% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastier.R
\docType{package}
\name{beastier}
\alias{beastier}
\alias{beastier-package}
\title{beastier: A package to call BEAST2.}
\description{
'beastier' allows to call BEAST2, a popular
Bayesian phylogenetics tool, using
an R interface. 'beastier' closely follows the interface
of BEAST2, including its default settings.
}
\seealso{
These are packages associated with 'beastier':
\itemize{
  \item{
    The package 'beautier' can create
    BEAST2 input files from R
  }
  \item{
    The package 'tracerer' can parse
    BEAST2 output files from R
  }
  \item{
    The package 'babette' combines the
    functionality of 'beautier',
    'beastier' and 'tracerer'
    into a single workflow
  }
}
}
