% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastier_report.R
\name{beastier_report}
\alias{beastier_report}
\title{Create a \link{beastier} report, to be used when reporting bugs}
\usage{
beastier_report()
}
\description{
Create a \link{beastier} report, to be used when reporting bugs
}
\author{
Richèl J.C. Bilderbeek
}
