% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastier_report.R
\name{beastier_report}
\alias{beastier_report}
\title{Create a \link{beastier} report, to be used when reporting bugs}
\usage{
beastier_report(
  beast2_folder = get_default_beast2_folder(),
  os = rappdirs::app_dir()$os
)
}
\arguments{
\item{beast2_folder}{the folder where the BEAST2 is installed.
Note that this is not the folder where the BEAST2 executable
is installed: the BEAST2 executable is in a subfolder.
Use \link{get_default_beast2_folder} to get the default BEAST2
folder.
Use \link{get_default_beast2_bin_path} to get the full path to
the default BEAST2 executable.}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\description{
Create a \link{beastier} report, to be used when reporting bugs
}
\author{
Richèl J.C. Bilderbeek
}
