% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_internal_filenames.R
\name{check_beast2_internal_filenames}
\alias{check_beast2_internal_filenames}
\title{Check if the \code{beast2_internal_filenames} is a valid BEAST2 options object.}
\usage{
check_beast2_internal_filenames(beast2_internal_filenames)
}
\value{
nothing
}
\description{
Calls \code{stop} if the BEAST2 option object is invalid
}
\examples{
 library(testthat)

 expect_silent(check_beast2_internal_filenames(create_beast2_internal_filenames()))

 # Must stop on nonsense
 expect_error(check_beast2_internal_filenames(beast2_internal_filenames = "nonsense"))
 expect_error(check_beast2_internal_filenames(beast2_internal_filenames = NULL))
 expect_error(check_beast2_internal_filenames(beast2_internal_filenames = NA))
}
\seealso{
Use \link{create_beast2_internal_filenames} to create a valid
  BEAST2 options object
}
\author{
Richèl J.C. Bilderbeek
}
