% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/check_beast2_options_do_not_overwrite_existing_files.R
\name{check_beast2_options_do_not_overwrite_existing_files}
\alias{check_beast2_options_do_not_overwrite_existing_files}
\title{Check if the \code{beast2_options} will not overwrite
existing files, when the 'overwrite' options is set to \code{FALSE}}
\usage{
check_beast2_options_do_not_overwrite_existing_files(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\description{
Will \link{stop} if a file is threatened to be overwritten
}
\author{
Richèl J.C. Bilderbeek
}
