% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_beast2_optionses.R
\name{check_beast2_optionses}
\alias{check_beast2_optionses}
\title{Check if the \code{beast2_options} is a valid BEAST2 options object.}
\usage{
check_beast2_optionses(beast2_optionses)
}
\arguments{
\item{beast2_optionses}{list of one or more \code{beast2_options}
structures,
as can be created by \link{create_beast2_options}.
Use of reduplicated plural to achieve difference with
\code{beast2_options}}
}
\value{
nothing
}
\description{
Calls \code{stop} if the BEAST2 option object is invalid
}
\examples{
 library(testthat)

 expect_silent(check_beast2_optionses(list(create_beast2_options())))

 # Must stop on nonsense
 expect_error(check_beast2_optionses("nonsense"))
 expect_error(check_beast2_optionses(NULL))
 expect_error(check_beast2_optionses(NA))
}
\seealso{
Use \link{create_beast2_options} to create a valid
  BEAST2 options object
}
\author{
Richèl J.C. Bilderbeek
}
