% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_input_filename.R
\name{check_input_filename_validity}
\alias{check_input_filename_validity}
\title{Checks the input filename.
Will stop if there is a problem with the input filename.}
\usage{
check_input_filename_validity(input_filename, beast2_path, verbose)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.}

\item{beast2_path}{name of either a BEAST2 binary file
(usually simply \code{beast})
or a BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_bin_path} to get
the default BEAST binary file's path
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\value{
nothing. Will call \code{\link{stop}} if the input file is invalid
}
\description{
Checks the input filename.
Will stop if there is a problem with the input filename.
}
\author{
Richel J.C. Bilderbeek
}
