% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_n_threads.R
\name{check_n_threads}
\alias{check_n_threads}
\title{Check if the input is a valid RNG seed.}
\usage{
check_n_threads(n_threads)
}
\arguments{
\item{n_threads}{the number of computational threads to use.
Use \link{NA} to use the BEAST2 default of 1.}
}
\description{
Will \link{stop} if not.
}
\author{
Richel J.C. Bilderbeek
}
