% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_os.R
\name{check_os}
\alias{check_os}
\title{Checks if the operating system is supported}
\usage{
check_os(os)
}
\arguments{
\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
nothing. Will stop if the OS is unsupported
}
\description{
Checks if the operating system is supported
}
\author{
Richèl J.C. Bilderbeek
}
