% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_options.R
\name{create_beast2_options}
\alias{create_beast2_options}
\title{Function to create a set of BEAST2 options.}
\usage{
create_beast2_options(output_log_filename = tempfile(fileext = "log"),
  output_trees_filenames = tempfile(fileext = ".trees"),
  output_state_filename = tempfile(fileext = ".xml.state"),
  rng_seed = NA, n_threads = NA, use_beagle = FALSE,
  overwrite = TRUE, beast2_path = get_default_beast2_path(),
  verbose = FALSE)
}
\arguments{
\item{output_log_filename}{name of the .log file to create}

\item{output_trees_filenames}{one or more names for .trees file to create.
There will be one .trees file created per alignment in the input
file. The number of alignments must equal the number of .trees
filenames, else an error is thrown. Alignments are sorted alphabetically
by their IDs}

\item{output_state_filename}{name of the .xml.state file to create}

\item{rng_seed}{the random number generator seed of the BEAST2 run.
Must be a non-zero positive integer value or \link{NA}.
If \code{rng_seed} is \link{NA}, BEAST2 will pick a random seed}

\item{n_threads}{the number of computational threads to use.
Use \link{NA} to use the BEAST2 default of 1.}

\item{use_beagle}{use BEAGLE if present}

\item{overwrite}{if TRUE: overwrite the \code{.log}
 and \code{.trees} files if one of these exists.
 If FALSE, BEAST2 will not be started if
 \itemize{
   \item{the \code{.log} file exists}
   \item{the \code{.trees} files exist}
   \item{the \code{.log} file created by BEAST2 exists}
   \item{the \code{.trees} files created by BEAST2 exist}
}}

\item{beast2_path}{name of either a BEAST2 binary file
(usually simply \code{beast})
or a BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_bin_path} to get
the default BEAST binary file's path
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\value{
a BEAST2 options structure
}
\description{
BEAUti options are settings that differ between BEAUti
version. The use of these options is mostly for testing
older versions
Whatever option chosen here, the created XML file will be valid.
}
\examples{
  beast2_options <- create_beast2_options()
}
\author{
Richel J.C. Bilderbeek
}
