% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_run_cmd.R
\name{create_beast2_run_cmd}
\alias{create_beast2_run_cmd}
\title{Creates the terminal command to run BEAST2}
\usage{
create_beast2_run_cmd(input_filename, output_state_filename,
  rng_seed = NA, n_threads = NA, use_beagle = FALSE,
  overwrite = FALSE, beast2_jar_path = get_default_beast2_jar_path())
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.}

\item{output_state_filename}{name of the BEAST2 output file that
stores the state
(usually has a \code{.xml.state} extension)}

\item{rng_seed}{the random number generator seed}

\item{n_threads}{number of threads to use}

\item{use_beagle}{use BEAGLE if present}

\item{overwrite}{if TRUE: overwrite the \code{.log}
 and \code{.trees} files if one of these exists.
 If FALSE, BEAST2 will not be started if
 \itemize{
   \item{the \code{.log} file exists}
   \item{the \code{.trees} files exist}
   \item{the \code{.log} file created by BEAST2 exists}
   \item{the \code{.trees} files created by BEAST2 exist}
}}

\item{beast2_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
one character string that calls the BEAST2
  \code{.jar} file, starting with \code{"java -jar "}
}
\description{
Creates the terminal command to run BEAST2
}
\examples{
  cmd <- create_beast2_run_cmd(
    input_filename = "input.xml",
    output_state_filename = "output.xml.state"
  )
  testit::assert(substr(cmd, 1, 10) == "java -jar ")
}
\author{
Richel J.C. Bilderbeek
}
