% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_validate_cmd.R
\name{create_beast2_validate_cmd}
\alias{create_beast2_validate_cmd}
\title{Creates the terminal command to validate a BEAST2 input file}
\usage{
create_beast2_validate_cmd(input_filename,
  beast2_jar_path = get_default_beast2_jar_path())
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.}

\item{beast2_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
a numeric vector, of which the first element
  is the command (\code{java}, in this case),
  and the others are arguments (\code{-jar}, in this case, followed
  by more arguments.
}
\description{
Creates the terminal command to validate a BEAST2 input file
}
\examples{
  cmds <- create_beast2_validate_cmd(
    input_filename = "input.xml"
  )
  testit::assert(cmds[1] == "java")
  testit::assert(cmds[2] == "-jar")
}
\author{
Richel J.C. Bilderbeek
}
