% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beast2_validate_cmd.R
\name{create_beast2_validate_cmd}
\alias{create_beast2_validate_cmd}
\title{Creates the terminal command to validate a BEAST2 input file}
\usage{
create_beast2_validate_cmd(input_filename,
  beast2_jar_path = get_default_beast2_jar_path())
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file.
This file usually has an \code{.xml} extension.}

\item{beast2_jar_path}{name of the BEAST2 jar file
(usually has a \code{.jar} extension).
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
one character string that calls the BEAST2
  \code{.jar} file, starting with \code{"java -jar "}
}
\description{
Creates the terminal command to validate a BEAST2 input file
}
\examples{
  cmd <- create_beast2_validate_cmd(
    input_filename = "input.xml"
  )
  testit::assert(substr(cmd, 1, 10) == "java -jar ")
}
\author{
Richel J.C. Bilderbeek
}
