% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beastier_tempfolder.R
\name{create_beastier_tempfolder}
\alias{create_beastier_tempfolder}
\title{Create the temporary folder as used by \link{beastier}}
\usage{
create_beastier_tempfolder()
}
\value{
nothing
}
\description{
Create the temporary folder as used by \link{beastier}
}
\examples{
check_empty_beaustier_folders()

create_beastier_tempfolder()

remove_beaustier_folders()
check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
