% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beautier_tempfolder.R
\name{create_beautier_tempfolder}
\alias{create_beautier_tempfolder}
\title{Deprecated function, use \link[beautier]{create_beautier_tempfolder}.}
\usage{
create_beautier_tempfolder()
}
\value{
nothing
}
\description{
Create the default `beautier` temporary folder
}
\author{
Richèl J.C. Bilderbeek
}
