% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beautier_tempfolder.R
\name{create_beautier_tempfolder}
\alias{create_beautier_tempfolder}
\title{Create the default `beautier` temporary folder}
\usage{
create_beautier_tempfolder()
}
\value{
nothing
}
\description{
Create the default `beautier` temporary folder
}
\examples{
create_beautier_tempfolder()

remove_beautier_folder()

check_empty_beautier_folder()
}
\author{
Richèl J.C. Bilderbeek
}
