% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_temp_input_filename.R
\name{create_temp_input_filename}
\alias{create_temp_input_filename}
\title{Create a temporary filename for the BEAST2 XML filename}
\usage{
create_temp_input_filename()
}
\description{
Create a temporary filename for the BEAST2 XML filename
}
