% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_param_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(clock_model, clock_models, crown_age, crown_ages,
  fasta_filename, fasta_filenames, fixed_crown_age, fixed_crown_ages,
  initial_phylogenies, input_fasta_filenames, mcmc, misc_options,
  output_xml_filename, sequence_length, site_model, site_models, tree_prior,
  tree_priors, verbose)
}
\arguments{
\item{clock_model}{a \code{beautier} clock model}

\item{clock_models}{a list of one or more \code{beautier} clock models}

\item{crown_age}{the crown age of the phylogeny}

\item{crown_ages}{the crown ages of the phylogenies. Set to NA
if the crown age needs to be estimated}

\item{fasta_filename}{a FASTA filename.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{fixed_crown_age}{determines if the phylogeny's crown age is
fixed. If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{fixed_crown_ages}{one or more booleans to determine if the
phylogenies' crown ages are fixed.
If FALSE, crown age is estimated by BEAST2. If TRUE,
the crown age is fixed to the crown age
of the initial phylogeny.}

\item{initial_phylogenies}{one or more MCMC chain's initial phylogenies.
Each one set to NA will result in BEAST2 using a random phylogeny. Else
the phylogeny is assumed to be of class ape::phylo.}

\item{input_fasta_filenames}{One or more FASTA filenames.
Use \code{\link{get_fasta_filename}} to obtain a testing FASTA filename.}

\item{mcmc}{one \code{beautier} mcmc object}

\item{misc_options}{one \code{beautier} misc_options object}

\item{output_xml_filename}{Name of the XML parameter file created by this
function. BEAST2 uses this file as input.}

\item{sequence_length}{a DNA sequence length, in base pairs}

\item{site_model}{a \code{beautier} site model}

\item{site_models}{one or more \code{beautier} site models}

\item{tree_prior}{a \code{beautier} tree prior}

\item{tree_priors}{one or more \code{beautier} tree priors}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\author{
Richel J.C. Bilderbeek
}
