% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beast2_example_filenames.R
\name{get_beast2_example_filenames}
\alias{get_beast2_example_filenames}
\title{Get a list with the full paths of all
BEAST2 example filenames}
\usage{

  get_beast2_example_filenames(beast2_folder = get_default_beast2_folder())
}
\arguments{
\item{beast2_folder}{the folder where the BEAST2 is installed.
Note that this is not the folder where the BEAST2 executable
is installed: the BEAST2 executable is in a subfolder.
Use \link{get_default_beast2_folder} to get the default BEAST2
folder.
Use \link{get_default_beast2_bin_path} to get the full path to
the default BEAST2 executable.}
}
\value{
a list with the full paths of all
  BEAST2 example filenames
}
\description{
Get a list with the full paths of all
BEAST2 example filenames
}
