% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_download_url.R
\name{get_default_beast2_download_url}
\alias{get_default_beast2_download_url}
\title{Get the default BEAST2 download URL,
which depends on the operating system}
\usage{
get_default_beast2_download_url(
  beast2_version = beastier::get_default_beast2_version(),
  os = rappdirs::app_dir()$os
)
}
\arguments{
\item{beast2_version}{the version of BEAST2. By
default, this is the version as returned by
\link{get_default_beast2_version}}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
the URL where BEAST2 can be downloaded from
}
\description{
Get the default BEAST2 download URL,
which depends on the operating system
}
\examples{
get_default_beast2_download_url()
}
\author{
Richèl J.C. Bilderbeek
}
