% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_download_url.R
\name{get_default_beast2_download_url_win}
\alias{get_default_beast2_download_url_win}
\title{Get the BEAST2 download URL for Windows}
\usage{
get_default_beast2_download_url_win(
  beast2_version = get_default_beast2_version()
)
}
\arguments{
\item{beast2_version}{the version of BEAST2. By
default, this is the version as returned by
\link{get_default_beast2_version}}
}
\value{
the URL where BEAST2 can be downloaded from
}
\description{
Get the BEAST2 download URL for Windows
}
\author{
Richèl J.C. Bilderbeek
}
