% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_beast2_jar_path.R
\name{get_default_beast2_jar_path}
\alias{get_default_beast2_jar_path}
\title{Get the default BEAST2 jar file's path}
\usage{
get_default_beast2_jar_path()
}
\description{
Get the default BEAST2 jar file's path
}
\examples{
  testit::assert(get_default_beast2_jar_path()
    == "~/Programs/beast/lib/beast.jar")
}
\seealso{
Use \link{get_default_beast2_folder} to get the default
  folder in which BEAST2 is installed
}
\author{
Richel J.C. Bilderbeek
}
