% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_java_path.R
\name{get_default_java_path}
\alias{get_default_java_path}
\title{Obtains the default path to the Java executable}
\usage{
get_default_java_path(os = rappdirs::app_dir()$os)
}
\arguments{
\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
the default path to the Java executable
}
\description{
Obtains the default path to the Java executable
}
\author{
Richel J.C. Bilderbeek
}
