% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trees_filenames.R
\name{get_trees_filenames}
\alias{get_trees_filenames}
\title{Get the .trees filenames that BEAST2 will produce}
\usage{
get_trees_filenames(input_filename)
}
\arguments{
\item{input_filename}{name of a BEAST2 XML file}
}
\value{
character vector with the names of the .trees files that BEAST2
  will produce
}
\description{
Get the .trees filenames that BEAST2 will produce
}
\author{
Richel J.C. Bilderbeek
}
