% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_beast2.R
\name{install_beast2}
\alias{install_beast2}
\title{Install BEAST2}
\usage{
install_beast2(folder_name = rappdirs::user_data_dir())
}
\arguments{
\item{folder_name}{name of the folder where the BEAST2 files will
be put. The name of the BEAST2 jar file will be at
\code{[folder_name]/beast/lib/beast.jar}}
}
\description{
Install BEAST2
}
\note{
This function only tested to work under GNU/Linux
}
\examples{
  install_beast2(folder_name = get_default_beast2_folder())
  beast2_jar_path <- paste0(folder_name, "/beast/lib/beast.jar")
  testit::assert(file.exists(beast2_jar_path))
}
\author{
Richel J.C. Bilderbeek
}
