% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_beast2.R
\name{install_beast2}
\alias{install_beast2}
\title{Install BEAST2}
\usage{
install_beast2(folder_name = rappdirs::user_data_dir(), verbose = TRUE)
}
\arguments{
\item{folder_name}{name of the folder where the BEAST2 files will
be put. The name of the BEAST2 jar file will be at
\code{[folder_name]/beast/lib/beast.jar}}

\item{verbose}{show debug info}
}
\description{
Install BEAST2
}
\note{
This function only tested to work under GNU/Linux
}
\examples{
  install_beast2()
  testit::assert(file.exists(get_default_beast2_jar_path()))
}
\author{
Richel J.C. Bilderbeek
}
