% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_beast2_input_file.R
\name{is_beast2_input_file}
\alias{is_beast2_input_file}
\title{Is a file a valid BEAST2 input file?}
\usage{
is_beast2_input_file(filename, show_warnings = FALSE, verbose = FALSE,
  beast2_jar_path = get_default_beast2_jar_path())
}
\arguments{
\item{filename}{name of the BEAST2 XML input file}

\item{show_warnings}{if TRUE, warnings will shown}

\item{verbose}{if TRUE, additional information is displayed, that
is potentially useful in debugging}

\item{beast2_jar_path}{the path of \code{beast.jar}.
Use \link{get_default_beast2_jar_path} to get
the default BEAST jar file's path}
}
\value{
TRUE if the file is valid, FALSE if not
}
\description{
Is a file a valid BEAST2 input file?
}
\seealso{
Use \code{\link{are_beast2_input_lines}} to check the lines
}
\author{
Richel J.C. Bilderbeek
}
