% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_appveyor}
\alias{is_on_appveyor}
\title{Determines if the environment is AppVeyor}
\usage{
is_on_appveyor()
}
\value{
TRUE if run on AppVeyor, FALSE otherwise
}
\description{
Determines if the environment is AppVeyor
}
\examples{
  if (is_on_appveyor()) {
    print("Running on AppVeyor")
  }
}
\author{
Richèl J.C. Bilderbeek
}
