% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_on_ci.R
\name{is_on_ci}
\alias{is_on_ci}
\title{Determines if the environment is Travis CI}
\usage{
is_on_ci()
}
\value{
TRUE if run on Travis CI, FALSE otherwise
}
\description{
Determines if the environment is Travis CI
}
\author{
Richèl J.C. Bilderbeek
}
