% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beast2_internal_filenames.R
\name{print_beast2_internal_filenames}
\alias{print_beast2_internal_filenames}
\title{Print a \code{beast2_internal_filenames} as a table}
\usage{
print_beast2_internal_filenames(beast2_internal_filenames)
}
\arguments{
\item{beast2_internal_filenames}{a list of internally
used BEAST2 filenames,
as created by \link{create_beast2_internal_filenames}}
}
\description{
Print a \code{beast2_internal_filenames} as a table
}
