% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beast2_options.R
\name{print_beast2_options}
\alias{print_beast2_options}
\title{Pretty-print a `beast2_options`}
\usage{
print_beast2_options(beast2_options)
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\value{
Nothing. Will display the `beast2_options` using \link{cat}.
}
\description{
Pretty-print a `beast2_options`
}
\examples{
print_beast2_options(create_beast2_options())
}
\author{
Richèl J.C. Bilderbeek
}
