% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_beast2_options.R
\name{print_beast2_options}
\alias{print_beast2_options}
\title{Pretty-print a \code{beast2_options}}
\usage{
print_beast2_options(beast2_options)
}
\description{
Pretty-print a \code{beast2_options}
}
