% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_beautier_folder.R
\name{remove_beautier_folder}
\alias{remove_beautier_folder}
\title{Deprecated function, use \link[beautier]{remove_beautier_folder}.}
\usage{
remove_beautier_folder()
}
\value{
Nothing.
}
\description{
Deprecated function, use \link[beautier]{remove_beautier_folder}.
}
\details{
Check there are no files in the default `beautier` folder.
The goal is to make sure no temporary files are left undeleted.
Will \link{stop} if there are files in the `beautier` folder.
}
\author{
Richèl J.C. Bilderbeek
}
