% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_beast2.R
\name{run_beast2}
\alias{run_beast2}
\title{Run BEAST2}
\usage{
run_beast2(input_filename, output_log_filename = "out.log",
  output_trees_filenames = get_trees_filenames(input_filename),
  output_state_filename = "out.xml.state",
  beast_jar_path = "~/Programs/beast/lib/beast.jar", verbose = FALSE)
}
\arguments{
\item{input_filename}{the name of a BEAST2 input XML file}

\item{output_log_filename}{name of the .log file to create}

\item{output_trees_filenames}{one or more names for .trees file to create.
There will be one .trees file created per alignment in the input
file. The number of alignments must equal the number of .trees
filenames, else an error is thrown. Alignments are sorted alphabetically
by their IDs}

\item{output_state_filename}{name of the .xml.state file to create}

\item{beast_jar_path}{the path of \code{beast.jar}}

\item{verbose}{show more (debug) output}
}
\description{
Run BEAST2
}
\examples{
  output_log_filename <- "out.log"
  output_trees_filename <- "out.trees"
  output_state_filename <- "out.state"

  run_beast2(
    input_filename = get_path("2_4.xml"),
    output_log_filename = output_log_filename,
    output_trees_filenames = output_trees_filename,
    output_state_filename = output_state_filename
  )

  testit::assert(file.exists(output_log_filename))
  testit::assert(file.exists(output_trees_filename))
  testit::assert(file.exists(output_state_filename))

}
\author{
Richel J.C. Bilderbeek
}
