% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_beast2_from_options.R
\name{run_beast2_from_options}
\alias{run_beast2_from_options}
\title{Run BEAST2}
\usage{
run_beast2_from_options(beast2_options = create_beast2_options())
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\description{
Run BEAST2
}
\examples{
  beast2_options <- create_beast2_options(
    output_log_filename = tempfile(fileext = ".log"),
    output_trees_filenames = tempfile(fileext = ".trees"),
    output_state_filename = tempfile(fileext = ".xml.state")
  )

  output <- run_beast2_from_options(
    input_filename = get_beastier_path("2_4.xml"),
    beast2_options = beast2_options
  )

  testit::assert(length(output) > 40)
  testit::assert(file.exists(beast2_options$output_log_filename))
  testit::assert(file.exists(beast2_options$output_trees_filenames))
  testit::assert(file.exists(beast2_options$output_state_filename))

}
\author{
Richel J.C. Bilderbeek
}
