% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_beast2_from_options.R
\name{run_beast2_from_options}
\alias{run_beast2_from_options}
\title{Run BEAST2}
\usage{
run_beast2_from_options(beast2_options = beastier::create_beast2_options())
}
\arguments{
\item{beast2_options}{a set of BEAST2 options,
that are the R equivalent of the BEAST2 command-line options,
as can be created by \link{create_beast2_options}}
}
\value{
The text sent to \code{STDOUT} and \code{STDERR}.
  It will create the file with name \code{output_state_filenames}
}
\description{
Run BEAST2
}
\examples{
check_empty_beaustier_folders()

if (is_beast2_installed() && beautier::is_on_ci()) {
  beast2_options <- create_beast2_options(
    input_filename = get_beastier_path("2_4.xml")
  )
  run_beast2_from_options(beast2_options)
  file.remove(beast2_options$output_state_filename)
  remove_beaustier_folders()
}
check_empty_beaustier_folders()
}
\author{
Richèl J.C. Bilderbeek
}
