% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_beast2.R
\name{uninstall_beast2}
\alias{uninstall_beast2}
\title{Uninstall BEAST2}
\usage{
uninstall_beast2(folder_name = get_default_beast2_folder())
}
\description{
Uninstall BEAST2
}
\author{
Richel J.C. Bilderbeek
}
