% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upgrade_beast2.R
\name{upgrade_beast2}
\alias{upgrade_beast2}
\title{Deprecated function to upgrade BEAST2.}
\usage{
upgrade_beast2(
  folder_name = rappdirs::user_data_dir(),
  os = rappdirs::app_dir()$os
)
}
\arguments{
\item{folder_name}{name of the folder where the BEAST2 files will
be put.
The name of the BEAST2 binary file will be at
\code{[folder_name]/beast/bin/beast}
The name of the BEAST2 jar file will be at
\code{[folder_name]/beast/lib/launcher.jar}}

\item{os}{name of the operating system,
must be \code{unix} (Linux, Mac) or \code{win} (Windows)}
}
\value{
Nothing.
A deprecation message using \link{stop} will be triggered
}
\description{
Deprecated function to upgrade BEAST2.
}
\author{
Richèl J.C. Bilderbeek
}
